improve: `ptrace_message` reset with `seccomp` (#191)
This commit improves how "ptrace_message" is cleaned by utilizing seccomp (Secure Computation Mode) to clear its value.

"ptrace_message" not being cleared is a Linux kernel vulnerability/bug that impacts all versions below 6.1, as the fix only came in 6.1, and because of that, we need to find ways to "0" it. This is the second fix for that, being the GKI2 fix the first: https://github.com/PerformanC/ReZygisk/commit/70697be9a5230b3e254a9df68b8cba599dbbab90